Refactor and prepare layout for shared work
