--- centrifuge-app/src/components/LayoutBase/config.ts
@@ -1,9 +1,0 @@
-export const config = {
-  HEADER_HEIGHT: 60,
-  TOOLBAR_HEIGHT: 50,
-  LAYOUT_MAX_WIDTH: 1800,
-  SIDEBAR_WIDTH: 80,
-  SIDEBAR_WIDTH_EXTENDED: 220,
-  PADDING_MAIN: [2, 2, 3, 3, 5],
-  WALLET_WIDTH: [200, 264],
-}
