+++ centrifuge-app/src/components/LayoutBase/index.tsx
@@ -1,4 +1,5 @@
 import { WalletMenu } from '@centrifuge/centrifuge-react'
+import { Box } from '@centrifuge/fabric'
 import * as React from 'react'
 import { Footer } from '../Footer'
 import { LoadBoundary } from '../LoadBoundary'
@@ -6,7 +7,6 @@ import { LogoLink } from '../LogoLink'
 import { Menu } from '../Menu'
 import { OnboardingStatus } from '../OnboardingStatus'
 import { SideDrawerProps } from '../SideDrawer'
-import { config } from './config'
 import {
   FooterContainer,
   HeaderBackground,
@@ -35,9 +35,9 @@ export function LayoutBase({ children, sideDrawer }: LayoutBaseProps) {
           <LogoLink />
         </LogoContainer>
 
-        <WalletContainer px={config.PADDING_MAIN}>
+        <WalletContainer px={[2, 2, 3, 3, 5]}>
           <WalletPositioner>
-            <WalletInner minWidth={config.WALLET_WIDTH}>
+            <WalletInner minWidth={[200, 264]}>
               <WalletMenu menuItems={[<OnboardingStatus />]} />
             </WalletInner>
           </WalletPositioner>
@@ -48,7 +48,9 @@ export function LayoutBase({ children, sideDrawer }: LayoutBaseProps) {
         </ToolbarContainer>
 
         <LoadBoundary>
-          <MainContainer as="main">{children}</MainContainer>
+          <MainContainer as="main">
+            <Box maxWidth={1800}>{children}</Box>
+          </MainContainer>
         </LoadBoundary>
 
         <FooterContainer>
