+++ centrifuge-app/src/components/LayoutBase/styles.tsx
@@ -1,11 +1,16 @@
 import { Box, Grid, Shelf, Stack } from '@centrifuge/fabric'
 import styled from 'styled-components'
-import { config } from './config'
 
 // the main breakpoint to switch from stacked to columns layout
 const BREAK_POINT_COLUMNS = 'M'
 // breakpoint from minimal to extended left sidebar width
 const BREAK_POINT_SIDEBAR_EXTENDED = 'L'
+
+const HEADER_HEIGHT = 60
+const TOOLBAR_HEIGHT = 50
+const SIDEBAR_WIDTH = 80
+const SIDEBAR_WIDTH_EXTENDED = 220
+const LAYOUT_MAX_WIDTH = 1800
 
 export const Root = styled(Box)`
   position: relative;
@@ -29,7 +34,7 @@ export const Inner = styled(Grid)`
   min-height: 100%;
 
   align-items: start;
-  grid-template-rows: 0px ${config.HEADER_HEIGHT}px 1fr auto ${config.TOOLBAR_HEIGHT}px;
+  grid-template-rows: 0px ${HEADER_HEIGHT}px 1fr auto ${TOOLBAR_HEIGHT}px;
   grid-template-columns: auto 1fr;
   grid-template-areas:
     'header-background header-background'
@@ -39,8 +44,8 @@ export const Inner = styled(Grid)`
     'toolbar toolbar';
 
   @media (min-width: ${({ theme }) => theme.breakpoints[BREAK_POINT_COLUMNS]}) {
-    grid-template-rows: ${config.HEADER_HEIGHT}px minmax(max-content, 1fr) auto;
-    grid-template-columns: ${config.SIDEBAR_WIDTH}px 1fr;
+    grid-template-rows: ${HEADER_HEIGHT}px minmax(max-content, 1fr) auto;
+    grid-template-columns: ${SIDEBAR_WIDTH}px 1fr;
     grid-template-areas:
       'logo wallet'
       'toolbar main'
@@ -48,7 +53,7 @@ export const Inner = styled(Grid)`
   }
 
   @media (min-width: ${({ theme }) => theme.breakpoints[BREAK_POINT_SIDEBAR_EXTENDED]}) {
-    grid-template-columns: ${config.SIDEBAR_WIDTH_EXTENDED}px 1fr;
+    grid-template-columns: ${SIDEBAR_WIDTH_EXTENDED}px 1fr;
   }
 `
 
@@ -60,7 +65,7 @@ export const HeaderBackground = styled(Box)`
 
   grid-area: header-background;
   width: 100%;
-  height: ${config.HEADER_HEIGHT}px;
+  height: ${HEADER_HEIGHT}px;
 
   background-color: ${({ theme }) => theme.colors.backgroundPrimary};
   border-bottom: ${({ theme }) => `1px solid ${theme.colors.borderSecondary}`};
@@ -76,13 +81,13 @@ export const ToolbarContainer = styled(Box)`
   position: sticky;
   left: 0;
   bottom: 0;
-  height: ${config.TOOLBAR_HEIGHT}px;
+  height: ${TOOLBAR_HEIGHT}px;
 
   border-top: ${({ theme }) => `1px solid ${theme.colors.borderSecondary}`};
   background-color: ${({ theme }) => theme.colors.backgroundPrimary};
 
   @media (min-width: ${({ theme }) => theme.breakpoints[BREAK_POINT_COLUMNS]}) {
-    top: ${({ theme }) => theme.space[4] + config.HEADER_HEIGHT}px;
+    top: ${({ theme }) => theme.space[4] + HEADER_HEIGHT}px;
     bottom: auto;
     height: auto;
 
@@ -101,7 +106,7 @@ export const LogoContainer = styled(Stack)`
   top: 0;
 
   grid-area: logo;
-  height: ${config.HEADER_HEIGHT}px;
+  height: ${HEADER_HEIGHT}px;
   padding-left: ${({ theme }) => theme.space[2]}px;
   justify-content: center;
 
@@ -122,13 +127,13 @@ export const WalletContainer = styled(Stack)`
 `
 
 export const WalletPositioner = styled(Shelf)`
-  max-width: ${config.LAYOUT_MAX_WIDTH}px;
+  max-width: ${LAYOUT_MAX_WIDTH}px;
   justify-content: flex-end;
   align-items: flex-start;
 `
 
 export const WalletInner = styled(Stack)`
-  height: ${config.HEADER_HEIGHT}px;
+  height: ${HEADER_HEIGHT}px;
   justify-content: center;
   pointer-events: auto; // resetting pointer events
 
@@ -142,7 +147,7 @@ export const MainContainer = styled(Box)`
   align-self: stretch;
 
   @media (min-width: ${({ theme }) => theme.breakpoints[BREAK_POINT_COLUMNS]}) {
-    margin-top: calc(${config.HEADER_HEIGHT}px * -1);
+    margin-top: calc(${HEADER_HEIGHT}px * -1);
     border-left: ${({ theme }) => `1px solid ${theme.colors.borderSecondary}`};
   }
 `
