+++ centrifuge-app/src/components/Portfolio/AssetAllocation.tsx
@@ -1,24 +1,26 @@
 import { useBalances } from '@centrifuge/centrifuge-react'
 import { Box, Text } from '@centrifuge/fabric'
 import * as React from 'react'
-import { useAddress } from '../utils/useAddress'
-import { AllocationCard } from './AllocationCard'
+import { useAddress } from '../../utils/useAddress'
 
-export function AddressAllocation() {
+export function AssetAllocation() {
   const address = useAddress()
   const balances = useBalances(address)
 
   return !!balances?.tranches && !!balances?.tranches.length ? (
-    <Box>
+    <Box as="article">
+      <Text as="h2" variant="heading2">
+        Allocation
+      </Text>
       <Box as="ul" role="list">
         {balances?.tranches.map((tranche, index) => (
           <Box as="li" key={`${tranche.trancheId}${index}`}>
-            <AllocationCard {...tranche} />
+            <Box>
+              <Text>Asset Class</Text>{' '}
+            </Box>
           </Box>
         ))}
       </Box>
     </Box>
-  ) : (
-    <Text>No data</Text>
-  )
+  ) : null
 }
