+++ centrifuge-app/src/pages/Portfolio/Transactions/index.tsx
@@ -1,33 +1,26 @@
 import { Box, Stack, Text } from '@centrifuge/fabric'
 import * as React from 'react'
-import { AddressTransactions } from '../../../components/AddressTransactions'
-import { PageWithSideBar } from '../../../components/PageWithSideBar'
+import { LayoutBase } from '../../../components/LayoutBase'
+import { Transactions } from '../../../components/Portfolio/Transactions'
 import { useAddress } from '../../../utils/useAddress'
 
 export function TransactionsPage() {
-  return (
-    <PageWithSideBar>
-      <Transactions />
-    </PageWithSideBar>
-  )
-}
-
-function Transactions() {
   const address = useAddress()
-
   return (
-    <Stack>
-      <Box as="header">
-        <Text as="h1" variant="heading1">
-          Transaction history
-        </Text>
-      </Box>
+    <LayoutBase>
+      <Stack>
+        <Box as="header">
+          <Text as="h1" variant="heading1">
+            Transaction history
+          </Text>
+        </Box>
 
-      {!!address ? (
-        <AddressTransactions />
-      ) : (
-        <Text as="strong">You need to connect your wallet to see your transactions</Text>
-      )}
-    </Stack>
+        {!!address ? (
+          <Transactions />
+        ) : (
+          <Text as="strong">You need to connect your wallet to see your transactions</Text>
+        )}
+      </Stack>
+    </LayoutBase>
   )
 }
