+++ centrifuge-app/src/pages/Portfolio/index.tsx
@@ -1,94 +1,69 @@
-import { useBalances } from '@centrifuge/centrifuge-react'
-import { Box, Card, Grid, Stack, Text } from '@centrifuge/fabric'
+import { Card, Grid, Stack, Text } from '@centrifuge/fabric'
 import * as React from 'react'
-import { Link } from 'react-router-dom'
-import { AddressAllocation } from '../../components/AddressAllocation'
-import { AddressTokens } from '../../components/AddressTokens'
-import { AddressTransactions } from '../../components/AddressTransactions'
-import { PageWithSideBar } from '../../components/PageWithSideBar'
+import { useTheme } from 'styled-components'
+import { LayoutBase } from '../../components/LayoutBase'
+import { BasePadding } from '../../components/LayoutBase/BasePadding'
 import { PoolList } from '../../components/PoolList'
-import { PortfolioRewards } from '../../components/PortfolioRewards'
+import { AssetAllocation } from '../../components/Portfolio/AssetAllocation'
+import { InvestedTokens } from '../../components/Portfolio/InvestedTokens'
+import { Rewards } from '../../components/Portfolio/Rewards'
+import { Transactions } from '../../components/Portfolio/Transactions'
 import { useAddress } from '../../utils/useAddress'
-import { useListedPools } from '../../utils/useListedPools'
 
 export function PortfolioPage() {
   return (
-    <PageWithSideBar>
+    <LayoutBase>
       <Portfolio />
-    </PageWithSideBar>
+    </LayoutBase>
   )
 }
 
 function Portfolio() {
   const address = useAddress()
-  const balances = useBalances(address)
-  const showFallback = !balances?.tranches.length && !balances?.currencies.length
+  const theme = useTheme()
 
+  if (!address) {
+    return (
+      <BasePadding>
+        <Text as="strong">You need to connect your wallet to see your portfolio</Text>
+      </BasePadding>
+    )
+  }
   return (
-    <Stack gap={4}>
-      <Stack as="header" gap={1}>
-        <Text as="h1" variant="heading1">
-          Your portfolio
-        </Text>
-        <Text as="p" variant="label1">
-          Track and manage your portfolio
-        </Text>
-      </Stack>
+    <>
+      <Stack gap={2}>
+        <BasePadding backgroundColor={theme.colors.backgroundSecondary} gap={4} pb={10}>
+          <Stack as="header" gap={1}>
+            <Text as="h1" variant="heading1">
+              Your portfolio
+            </Text>
+            <Text as="p" variant="label1">
+              Track and manage your portfolio
+            </Text>
+          </Stack>
 
-      {!!address ? (
-        <>
           <Grid gridTemplateColumns={['1.5fr 1fr']} gap={4}>
-            <Card as="article">
-              <Text as="h2">Portfolio stats</Text>
+            <Card as="article" p={1}>
+              <Text as="h2">Portfolio overview</Text>
             </Card>
 
-            <PortfolioRewards />
+            <Rewards />
           </Grid>
-
-          {showFallback ? (
-            <Fallback />
-          ) : (
-            <>
-              <Box as="article">
-                <Text as="h2" variant="heading2">
-                  Token overview
-                </Text>
-                <AddressTokens />
-              </Box>
-
-              <Stack as="article" gap={2}>
-                <Text as="h2" variant="heading2">
-                  Transaction history
-                </Text>
-                <AddressTransactions count={3} />
-                <Link to="portfolio/transactions">View all</Link>
-              </Stack>
-
-              <Box as="article">
-                <Text as="h2" variant="heading2">
-                  Allocation
-                </Text>
-                <AddressAllocation />
-              </Box>
-            </>
-          )}
-        </>
-      ) : (
-        <Text as="strong">You need to connect your wallet to see your portfolio</Text>
-      )}
-    </Stack>
-  )
-}
-
-function Fallback() {
-  const [listedPools, , metadataIsLoading] = useListedPools()
-
-  return (
-    <Stack as="article">
-      <Text as="h2" variant="heading2">
-        Explore opportunities
-      </Text>
-      <PoolList pools={listedPools.filter(({ reserve }) => reserve.max.toFloat() > 0)} isLoading={metadataIsLoading} />
-    </Stack>
+        </BasePadding>
+        <BasePadding>
+          <InvestedTokens />
+          <Transactions count={3} />
+          <AssetAllocation />
+        </BasePadding>
+        <BasePadding>
+          <Stack as="article" gap={2}>
+            <Text as="h2" variant="heading2">
+              Explore opportunities
+            </Text>
+            <PoolList />
+          </Stack>
+        </BasePadding>
+      </Stack>
+    </>
   )
 }
